W127 mdlFile triage: bank FUN_0034c890 nd82->11 (dropped u64 cast at FUN_003257e0 loop call); revert FUN_00332880 volatile staged-index probe (nd59->64) and FUN_003398c0 no-gain store rewrite. mdlFile 393->395 MATCH, over-window 149->147

+++ src/Yajima/y_runtime.c
@@ -12655,7 +12655,7 @@ void FUN_0044a240(float param_1,float param_2,float param_3,float param_4,int pa
 {
   u8 uVar3;
   u32 uVar4;
-  u32 uVar2;
+  u64 uVar2;
   int iVar1;
   
   uVar2 = FUN_001158b0(0,*(u32 *)(*(int *)(param_5 + 0x3c) + 4),param_6);
